filling of production and category tables
